Move tokenizer code into separate file.
