+++ renpy/lexer.py
@@ -22,18 +22,23 @@
 
 from __future__ import annotations
 
-import enum
 from typing import Any, Iterator, NamedTuple
 
 import os
-import io
 import re
-import sys
 import tokenize
 import contextlib
 import keyword
 
 import renpy
+
+from renpy.tokenizer import (
+    Line,
+    Token,
+    TokenKind,
+    TokenExactKind,
+    Tokenizer,
+)
 
 # The filename that's in the line text cache.
 line_text_filename = ""
@@ -51,7 +56,7 @@ def get_line_text(filename, lineno):
     global line_text_filename
     global line_text_cache
 
-    full_filename = renpy.exports.unelide_filename(filename)
+    full_filename = unelide_filename(filename)
 
     if full_filename != line_text_filename:
 
@@ -219,816 +224,6 @@ def unelide_filename(fn):
 # The filename that the start and end positions are relative to.
 original_filename = ""
 
-_DEBUG_TOKENIZATION = "RENPY_DEBUG_TOKENIZATION" in os.environ
-
-
-class Line:
-    """
-    Represents a logical line in a file, which is a sequence of tokens.
-    Line should ends with a NEWLINE token.
-    """
-
-    __slots__ = ['tokens', 'indent_depth', '_text']
-
-    def __init__(self, token: Token, *tokens: Token, indent_depth: int):
-        self.tokens: tuple[Token, ...] = (token, *tokens)
-        """
-        List of tokens that make up this line.
-        """
-
-        self.indent_depth: int = indent_depth
-        """
-        Indentation depth of this line.
-        """
-
-        self._text = None
-
-        if not _DEBUG_TOKENIZATION:
-            return
-
-        has_newline = False
-        physical_offset = token.physical_offset
-        prev_row = -1
-        prev_col = -1
-        tokens_iter = iter(self.tokens)
-        for t in tokens_iter:
-            # If there is a change in physical offset in between tokens, it
-            # is impossible to locate the line in the file.
-            if t.physical_offset != physical_offset:
-                raise ValueError("Tokens in logical line must be contiguous.")
-
-            if t.lineno < prev_row or t.lineno == prev_row and t.col_offset < prev_col:
-                raise ValueError(f"start ({t.lineno}, {t.lineno}) precedes "
-                                 f"previous end ({prev_row}, {prev_col})")
-
-            prev_row = t.end_lineno
-            prev_col = t.end_col_offset
-
-            match t.kind:
-                case TokenKind.INDENT | TokenKind.DEDENT:
-                    raise ValueError("Line can't contain INDENT or DEDENT tokens.")
-
-                case TokenKind.NEWLINE:
-                    has_newline = True
-                    break
-
-        if list(tokens_iter):
-            raise ValueError("NEWLINE must be the last token.")
-
-        if not has_newline:
-            raise ValueError("Line must end with NEWLINE token.")
-
-    @property
-    def filename(self) -> str:
-        """
-        Elided filename where line is located.
-        """
-
-        return self.tokens[0].filename
-
-    @property
-    def start(self) -> tuple[int, int]:
-        """
-        Linenumber, col offset of start of logical line.
-        """
-
-        return (self.tokens[0].lineno, self.tokens[0].col_offset)
-
-    @property
-    def end(self) -> tuple[int, int]:
-        """
-        Linenumber, col offset of end of logical line.
-        """
-
-        return (self.tokens[-1].end_lineno, self.tokens[-1].end_col_offset)
-
-    @property
-    def text(self) -> str:
-        """
-        Full text of logical line including whitespaces and comments.
-        """
-
-        if self._text is None:
-            prev_row = self.tokens[0].lineno
-            prev_col = 0
-            cont_line = False
-            result: list[str] = []
-
-            for t in self.tokens:
-                # Don't add spurious spaces before new line.
-                if t.kind in (TokenKind.NL, TokenKind.NEWLINE):
-                    cont_line = False
-                    prev_row += 1
-                    prev_col = 0
-                    continue
-
-                if row_offset := t.lineno - prev_row:
-                    if cont_line:
-                        # Assume people continue lines with single space before slash.
-                        result.append(" \\")
-
-                    result.append("\n" * row_offset)
-                    prev_col = 0
-
-                if col_offset := t.col_offset - prev_col:
-                    result.append(" " * col_offset)
-
-                result.append(t.string)
-                prev_row = t.end_lineno
-                prev_col = t.end_col_offset
-                cont_line = True
-
-            self._text = "".join(result)
-
-        return self._text
-
-    def __repr__(self):
-        return f"<Line {self.filename} {self.start}-{self.end}>"
-
-
-class TokenKind(enum.StrEnum):
-    """
-    Enum of all possible `Token.name` values.
-
-    Equality can be checked as lower-case string value.
-    """
-
-    # Special tokens.
-    INDENT = enum.auto()
-    DEDENT = enum.auto()
-    COMMENT = enum.auto()
-
-    # Non-terminating and terminating new lines.
-    # This is always \n in Ren'Py.
-    NL = enum.auto()
-    NEWLINE = enum.auto()
-
-    # Names.
-    NAME = enum.auto()
-
-    # Numbers.
-    NUMBER = enum.auto()
-
-    # Strings.
-    STRING = enum.auto()
-
-    # Operators.
-    OP = enum.auto()
-
-
-class TokenExactKind(enum.StrEnum):
-    """
-    Enum of all possible `Token.exact_name` values.
-
-    Equality can be checked as lower-case string value.
-    """
-
-    # Special tokens.
-    INDENT = enum.auto()
-    DEDENT = enum.auto()
-    COMMENT = enum.auto()
-
-    # Non-terminating and terminating new lines.
-    # This is always \n in Ren'Py.
-    NL = enum.auto()
-    NEWLINE = enum.auto()
-
-    # Names.
-    NAME = enum.auto()
-    KEYWORD = enum.auto()
-    IDENTIFIER = enum.auto()
-
-    # Numbers.
-    HEX = enum.auto()
-    BINARY = enum.auto()
-    OCTAL = enum.auto()
-    IMAG = enum.auto()
-    FLOAT = enum.auto()
-    INT = enum.auto()
-
-    # Strings.
-    STRING = enum.auto()
-    BYTES = enum.auto()
-    F_STRING = enum.auto()
-    RAW_TRIPLE_STRING = enum.auto()
-    TRIPLE_STRING = enum.auto()
-    RAW_STRING = enum.auto()
-
-    # Operators.
-    DOLLAR = enum.auto()
-    LPAR = enum.auto()
-    RPAR = enum.auto()
-    LSQB = enum.auto()
-    RSQB = enum.auto()
-    COLON = enum.auto()
-    COMMA = enum.auto()
-    SEMI = enum.auto()
-    PLUS = enum.auto()
-    MINUS = enum.auto()
-    STAR = enum.auto()
-    SLASH = enum.auto()
-    VBAR = enum.auto()
-    AMPER = enum.auto()
-    LESS = enum.auto()
-    GREATER = enum.auto()
-    EQUAL = enum.auto()
-    DOT = enum.auto()
-    PERCENT = enum.auto()
-    LBRACE = enum.auto()
-    RBRACE = enum.auto()
-    EQEQUAL = enum.auto()
-    NOTEQUAL = enum.auto()
-    LESSEQUAL = enum.auto()
-    GREATEREQUAL = enum.auto()
-    TILDE = enum.auto()
-    CIRCUMFLEX = enum.auto()
-    LEFTSHIFT = enum.auto()
-    RIGHTSHIFT = enum.auto()
-    DOUBLESTAR = enum.auto()
-    PLUSEQUAL = enum.auto()
-    MINEQUAL = enum.auto()
-    STAREQUAL = enum.auto()
-    SLASHEQUAL = enum.auto()
-    PERCENTEQUAL = enum.auto()
-    AMPEREQUAL = enum.auto()
-    VBAREQUAL = enum.auto()
-    CIRCUMFLEXEQUAL = enum.auto()
-    LEFTSHIFTEQUAL = enum.auto()
-    RIGHTSHIFTEQUAL = enum.auto()
-    DOUBLESTAREQUAL = enum.auto()
-    DOUBLESLASH = enum.auto()
-    DOUBLESLASHEQUAL = enum.auto()
-    AT = enum.auto()
-    ATEQUAL = enum.auto()
-    RARROW = enum.auto()
-    ELLIPSIS = enum.auto()
-    COLONEQUAL = enum.auto()
-
-
-class Token:
-    """
-    Represents a token in a file.
-    """
-
-    __slots__ = [
-        'kind',
-        'exact_kind',
-        'string',
-        'munged',
-        'filename',
-        'lineno',
-        'col_offset',
-        'end_lineno',
-        'end_col_offset',
-        'physical_offset',
-    ]
-
-    def __init__(
-        self,
-        type: int,
-        string: str,
-        filename: str,
-        lineno: int,
-        col_offset: int,
-        end_lineno: int,
-        end_col_offset: int,
-        physical_offset: tuple[int, int],
-    ):
-        match type:
-            case tokenize.ERRORTOKEN if string == "$":  # FIXME: In 3.12 it is OP
-                kind = TokenKind.OP
-                exact_kind = TokenExactKind.DOLLAR
-            case tokenize.ERRORTOKEN:
-                raise SyntaxError(f"Error token with value: {string!r} on line "
-                                  f"{lineno + physical_offset[0]}")
-
-            # Exact name is one of predefined.
-            case tokenize.OP:
-                kind = TokenKind.OP
-                type = tokenize.EXACT_TOKEN_TYPES[string]
-                exact_kind = TokenExactKind(tokenize.tok_name[type].lower())
-
-            # f-strings and bytes are expressions in Ren'Py,
-            # and raw strings are handled differently.
-            case tokenize.STRING:
-                kind = TokenKind.STRING
-
-                prefix = re.match(
-                    r'([urfbURFB])*("""|\'\'\'|"|\')', string)
-                assert prefix is not None
-                mods = prefix.group(1) or ""
-                quotes = prefix.group(2)
-
-                if "b" in mods or "B" in mods:
-                    exact_kind = TokenExactKind.BYTES
-                elif "f" in mods or "F" in mods:
-                    exact_kind = TokenExactKind.F_STRING
-                elif "r" in mods or "R" in mods:
-                    if quotes == "'''" or quotes == '"""':
-                        exact_kind = TokenExactKind.RAW_TRIPLE_STRING
-                    else:
-                        exact_kind = TokenExactKind.RAW_STRING
-                elif quotes == "'''" or quotes == '"""':
-                    exact_kind = TokenExactKind.TRIPLE_STRING
-                else:
-                    exact_kind = TokenExactKind.STRING
-
-            # Split numbers into complex, float, and integer.
-            case tokenize.NUMBER:
-                kind = TokenKind.NUMBER
-                if string.startswith("0x"):
-                    exact_kind = TokenExactKind.HEX
-                elif string.startswith("0b"):
-                    exact_kind = TokenExactKind.BINARY
-                elif string.startswith("0o"):
-                    exact_kind = TokenExactKind.OCTAL
-                elif string[-1] in "jJ":
-                    exact_kind = TokenExactKind.IMAG
-                else:
-                    symbols = set(string)
-                    # Definitely a float.
-                    if "." in symbols:
-                        exact_kind = TokenExactKind.FLOAT
-
-                    # 00e0 is a float and can be interpreted as hex,
-                    # but we assume this is a float.
-                    elif symbols.intersection("eE"):
-                        exact_kind = TokenExactKind.FLOAT
-
-                    else:
-                        exact_kind = TokenExactKind.INT
-
-            # Names can be keywords, soft-keywords (including Ren'Py) or
-            # identifiers. Otherwise it is a 0name, which is not an identifier,
-            # but e.g. is a valid attribute name.
-            case tokenize.NAME:
-                kind = TokenKind.NAME
-
-                if keyword.iskeyword(string):
-                    exact_kind = TokenExactKind.KEYWORD
-                elif string.isidentifier():
-                    exact_kind = TokenExactKind.IDENTIFIER
-                else:
-                    exact_kind = TokenExactKind.NAME
-
-            case _:
-                kind = TokenKind(tokenize.tok_name[type].lower())
-                exact_kind = TokenExactKind(tokenize.tok_name[type].lower())
-
-        self.kind: TokenKind = kind
-        """
-        Kind of the token, such as TokenKind.NAME, TokenKind.OP or TokenKind.NUMBER.
-        Can be iterpreted as a corresponding lower-cased string.
-        """
-
-        self.exact_kind: TokenExactKind = exact_kind
-        """
-        Exact kind of the token, such as TokenExactKind.KEYWORD, TokenExactKind.DOLLAR or TokenExactKind.INT.
-        Can be iterpreted as a corresponding upper-cased string.
-        """
-
-        self.string: str = string
-        """
-        String value of the token.
-        """
-
-        self.munged: str = string
-        """
-        Munged string of the token, if applicable.
-        """
-
-        self.filename: str = filename
-        """
-        Elided filename where token is located.
-        """
-
-        # Positions of the token in the source.
-        self.lineno: int = lineno
-        self.col_offset: int = col_offset
-        self.end_lineno: int = end_lineno
-        self.end_col_offset: int = end_col_offset
-
-        # Offset from logical positions to physical positions.
-        self.physical_offset: tuple[int, int] = physical_offset
-
-    def __repr__(self):
-        string = repr(self.string)
-        if len(string) > 30:
-            string = string[:15] + "..." + string[-15:]
-
-        location = f"{self.lineno}:{self.col_offset}-{self.end_lineno}:{self.end_col_offset}"
-        return f"<Token {self.exact_kind._name_!r} {string} {self.filename} {location}>"
-
-
-class Tokenizer:
-    """
-    This class is used to read files and strings as RenPy source code.
-
-    First, this class reads passed buffer one line at a time, saving normalized
-    physical lines.
-
-    List of normalization and checks that this class performs during file read:
-        1. Strips the BOM from the start of the file.
-        2. Converts _ren.py files to equivalent .rpy files.
-        3. Normalizes line endings to \n.
-        4. Disallows \t and \f characters.
-
-    Then, it tokenizes read lines saving result tokens into a list.
-    Tokenization works the same as Python `tokenize.generate_tokens` function
-    with the following additional normalizations and checks:
-        1. It produces Token objects instead of tuples.
-        2. It allow 0-prefix non-zero numbers (e.g. 001).
-        3. It allows for names to start with numbers (e.g. 1foo).
-        4. It disallows non-ASCII characters in names.
-        5. It munges names that starts with double underscores
-           (e.g. "__foo" -> f"_m1_{filename}__foo").
-        6. It also munges words inside strings that start with double
-           underscores.
-        7. It checks for non-terminated strings and parenthesis.
-        8. It checks for parenthesis order of closing.
-
-    After file is tokenized, it can be used to list logical lines and group them
-    into indented blocks.
-    """
-
-    @staticmethod
-    def _strip_bom(lines: Iterator[str]) -> Iterator[str]:
-        first = next(lines)
-        if first.startswith('\ufeff'):
-            yield first[1:]
-        else:
-            yield first
-
-        yield from lines
-
-    def _ren_py_to_rpy(self, filename: str, lines: Iterator[str]) -> Iterator[str]:
-        """
-        Takes an iterator of lines from a _ren.py file, and returns an iterator
-        of lines of equivalent .rpy file. This should retain line numbers.
-        """
-
-        # The prefix prepended to Python lines.
-        prefix = ""
-
-        # Col offset of non-prefixed lines.
-        col_offset = self.col_offset
-
-        # Possible states.
-        IGNORE = 0
-        RENPY = 1
-        PYTHON = 2
-
-        # The state the state machine is in.
-        state = IGNORE
-
-        open_linenumber = 0
-
-        for linenumber, l in enumerate(lines, start=1):
-            if state != RENPY:
-                if l.startswith('"""renpy'):
-                    state = RENPY
-                    open_linenumber = linenumber
-                    self.col_offset = col_offset
-                    yield "\n"
-                    continue
-
-            if state == RENPY:
-                if l.strip() == '"""':
-                    state = PYTHON
-                    yield "\n"
-                    continue
-
-                # Ignore empty and comments.
-                sl = l.strip()
-                if not sl:
-                    yield l
-                    continue
-
-                if sl[0] == "#":
-                    yield l
-                    continue
-
-                # Determine the prefix.
-                prefix = ""
-                for i in l:
-                    if i != ' ':
-                        break
-                    prefix += ' '
-
-                # If the line ends in ":", add 4 spaces to the prefix.
-                # XXX: This does not work for 'init python: # Comment'...
-                if sl[-1] == ":":
-                    prefix += "    "
-
-                # Deduct the column offset, so error positions will be correct
-                # for python code.
-                self.col_offset -= len(prefix)
-
-                yield l
-                continue
-
-            if state == PYTHON:
-                if l == "\n":
-                    yield l
-                else:
-                    yield f"{prefix}{l}"
-                continue
-
-            if state == IGNORE:
-                yield "\n"
-                continue
-
-        if self.no_errors:
-            return
-
-        if state == IGNORE:
-            raise SyntaxError(f'In {filename!r}, there are no """renpy blocks, so every line is ignored.')
-
-        if state == RENPY:
-            raise SyntaxError(f'In {filename!r}, there is a """renpy block '
-                              f'at line {open_linenumber} that is not terminated by """.')
-
-    @staticmethod
-    def _disallow_bad_whitespaces(lines: Iterator[str]) -> Iterator[str]:
-        for l in lines:
-            if "\t" in l:
-                raise SyntaxError(f"Tab character is not allowed in Ren'Py scripts.")
-
-            if "\f" in l:
-                raise SyntaxError(f"Form feed character is not allowed in Ren'Py scripts.")
-
-            yield l
-
-    @classmethod
-    def from_string(
-        cls,
-        filename: str,
-        code: str, *,
-        lineno=0,
-        col_offset=0,
-        no_errors=False,
-    ):
-        return cls(filename, io.StringIO(code, newline=None),
-                   lineno=lineno, col_offset=col_offset, no_errors=no_errors)
-
-    @classmethod
-    def from_file(
-        cls,
-        filename: str, *,
-        no_errors=False,
-    ):
-        if os.path.isabs(filename):
-            fn = filename
-        else:
-            fn = unelide_filename(filename)
-
-        return cls(filename, open(fn, encoding="utf-8", errors="python_strict", newline=None),
-                   no_errors=no_errors)
-
-    def __init__(
-        self,
-        filename: str,
-        lines: Iterator[str], *,
-        lineno=0,
-        col_offset=0,
-        no_errors=False,
-    ):
-        from collections import deque
-        self.original_filename: str = filename
-        self.filename: str = elide_filename(filename)
-
-        self.linenumber: int = lineno
-        self.col_offset: int = col_offset
-
-        self.no_errors: bool = no_errors
-
-        lines = self._strip_bom(lines)
-
-        if filename.endswith("_ren.py"):
-            lines = self._ren_py_to_rpy(filename, lines)
-
-        if not no_errors:
-            lines = self._disallow_bad_whitespaces(lines)
-
-        # List of tokenized, normalized physical lines.
-        self._physical_lines: list[str] = []
-
-        # This is used to buffer lines that are read but not tokenized,
-        # so one can list physical lines without more expensive tokenization
-        # pass.
-        self._next_lines_buffer: deque[str] = deque()
-
-        # Generator of lines that are yet to be read.
-        self._next_lines = lines
-
-        # List of result tokens.
-        self._tokens: list[Token] = []
-
-        self._tokenized = None
-
-    def _readline(self):
-        if self._next_lines_buffer:
-            line = self._next_lines_buffer.popleft()
-        else:
-            line = next(self._next_lines)
-
-        self._physical_lines.append(line)
-        return line
-
-    def _tokenize(self) -> Iterator[Token]:
-        hold_number = None
-        try:
-            for t in tokenize.generate_tokens(self._readline):
-                if t.type == tokenize.ENDMARKER:
-                    continue
-
-                start_lineno, start_colno = t.start
-                end_lineno, end_colno = t.end
-
-                result = Token(
-                    t.type,
-                    t.string,
-                    self.filename,
-                    start_lineno,
-                    start_colno,
-                    end_lineno,
-                    end_colno,
-                    (self.linenumber, self.col_offset),
-                )
-
-                if hold_number is None and t.type == tokenize.NUMBER:
-                    hold_number = result
-                    continue
-
-                # If we have '0' on hold check if we can collapse it
-                # with next token.
-                if hold_number is not None:
-                    # Tokens are not next to each other, yield both.
-                    if (
-                        hold_number.end_col_offset != result.col_offset or
-                        hold_number.end_lineno != result.lineno
-                    ):
-                        yield hold_number
-                        hold_number = None
-
-                    # Python 2 style octal numbers are still valid in Ren'Py.
-                    # Tokenize report this as two numbers next to each other.
-                    # This is no longer needed in Python 3.12+
-                    elif t.type == tokenize.NUMBER:
-                        hold_number.string += result.string
-                        hold_number.munged = hold_number.string
-                        hold_number.end_lineno = result.end_lineno
-                        hold_number.end_col_offset = result.end_col_offset
-
-                        # We don't yield it yet, because '01attr' is 3 Python tokens.
-                        continue
-
-                    # 0-prefixed name is a valid in Ren'Py (e.g. image attributes).
-                    # NUMBER "$1" and NAME "$2" should appear as NAME "$1$2"
-                    elif t.type == tokenize.NAME:
-                        result.string = f"{hold_number.string}{result.string}"
-                        result.munged = result.string
-                        result.lineno = hold_number.lineno
-                        result.col_offset = hold_number.col_offset
-                        hold_number = None
-
-                    # TODO: ERRORTOKEN "`" ... ERRORTOKEN "`" -> STRING???
-
-                    # Otherwise we yield a hold number and the result.
-                    else:
-                        yield hold_number
-                        hold_number = None
-
-                yield result
-
-        except IndentationError as err:
-            line, column = err.args[1][1:3]
-            line += self.linenumber
-            column += self.col_offset
-            if not self.no_errors:
-                raise SyntaxError(f"Indentation error in {self.filename}:{line}:{column}") from None
-
-        except tokenize.TokenError as err:
-            line, column = err.args[1]
-            line += self.linenumber
-            column += self.col_offset
-            if not self.no_errors:
-                raise SyntaxError(f"{err.args[0]} at {self.filename}:{line}:{column}") from None
-
-    def _yield_tokens(self) -> Iterator[Token]:
-        if self._tokenized:
-            yield from self._tokens
-            return
-
-        if self._tokenized is not None:
-            raise RuntimeError("_yield_tokens() called twice.")
-
-        self._tokenized = False
-
-        def unmatched_paren():
-            if not self.no_errors:
-                lineno = token.lineno + token.physical_offset[0]
-                col_offset = token.col_offset + token.physical_offset[1]
-                raise SyntaxError(f"unmatched '{token.string}' at {self.filename}:{lineno}:{col_offset}")
-
-        def wrong_paren():
-            if not self.no_errors:
-                lineno = token.lineno + token.physical_offset[0]
-                col_offset = token.col_offset + token.physical_offset[1]
-                raise SyntaxError(f"closing parenthesis '{parens[-1].string}' "
-                                  f"does not match opening parenthesis '{token.string}'"
-                                  f" at {self.filename}:{lineno}:{col_offset}")
-
-        prefix: str = munge_filename(self.filename)
-
-        munge_regexp = re.compile(r'\b__(\w+)')
-
-        def munge_string(m: re.Match[str]):
-            g1 = m.group(1)
-
-            if "__" in g1:
-                return m.group(0)
-
-            if g1.startswith("_"):
-                return m.group(0)
-
-            return prefix + m.group(1)
-
-        parens: list[Token] = []
-        for token in self._tokenize():
-            match token.exact_kind:
-                case TokenExactKind.RPAR | TokenExactKind.RBRACE | TokenExactKind.RSQB if not parens:
-                    unmatched_paren()
-                case TokenExactKind.RPAR if parens[-1].exact_kind != TokenExactKind.LPAR:
-                    wrong_paren()
-                case TokenExactKind.RBRACE if parens[-1].exact_kind != TokenExactKind.LBRACE:
-                    wrong_paren()
-                case TokenExactKind.RSQB if parens[-1].exact_kind != TokenExactKind.LSQB:
-                    wrong_paren()
-
-                case TokenExactKind.LPAR | TokenExactKind.LBRACE | TokenExactKind.LSQB:
-                    parens.append(token)
-                case TokenExactKind.RPAR | TokenExactKind.RBRACE | TokenExactKind.RSQB:
-                    parens.pop()
-
-            match token.kind:
-                case TokenKind.NAME | TokenKind.STRING if "__" in token.string:
-                    token.munged = munge_regexp.sub(
-                        munge_string, token.munged)
-
-                    token.munged = sys.intern(token.munged)
-
-            self._tokens.append(token)
-            yield token
-
-        # self._tokenize will raise SyntaxError for unclosed paren or string literal.
-        self._tokenized = True
-
-    def list_physical_lines(self) -> list[str]:
-        """
-        Return a list of physical lines of tokenized code, such that
-        `"".join(physical lines)` is the lines that tokenizer sees.
-        """
-
-        rv = [*self._physical_lines, *self._next_lines_buffer]
-        while True:
-            try:
-                line = next(self._next_lines)
-                self._next_lines_buffer.append(line)
-                rv.append(line)
-            except StopIteration:
-                return rv
-
-    def list_logical_lines(self) -> list[Line]:
-        """
-        Return a list of logical lines of tokenized code.
-        """
-
-        rv: list[Line] = []
-
-        depth = 0
-        tokens: list[Token] = []
-
-        for token in self._yield_tokens():
-            match token.kind:
-                # Empty lines and bare comments are not logical lines.
-                case TokenKind.NL | TokenKind.COMMENT if not tokens:
-                    continue
-                case TokenKind.INDENT:
-                    depth += 1
-                case TokenKind.DEDENT:
-                    depth -= 1
-                case _:
-                    tokens.append(token)
-
-            if token.kind is TokenKind.NEWLINE:
-                rv.append(Line(*tokens, indent_depth=depth))
-                tokens.clear()
-
-        return rv
-
-    def list_tokens(self) -> list[Token]:
-        return list(self._yield_tokens())
 
 
 word_regexp = r'[a-zA-Z_\u00a0-\ufffd][0-9a-zA-Z_\u00a0-\ufffd]*'
